Add aggregates endpoint
This patch adds a new `aggregates` endpoint that retrieves the
missioncontrol data from presto provided a list of dimensions, measurements and conditions.
This is done using an instance of QueryBuilder, which holds the logic
to create a sqlalchemy selectable object and runs the query. This is
the first iteration so it's limited in functionalities. At the moment
it's not possible to:
 - filter by a non-exact match
 - filter by a time range
 - use a custom temporal resolution