Don't update graphs unnecessarily in detail view (#172)
We were previously updating the measure/time graphs in the detail view
every time any aspect of the global state changed, even when such a
state did not effect the graphs at all (e.g. the user was just changing
the time picker). This operation could be extremely slow when we're
visualizing a large amount of data. Use react's shouldComponentUpdate
functionality to avoid unnecessary updates.