Fix monkeypatching our presto raw_query method
The new experiments etl caused the method that queries presto to be
imported earlier in the unit test process, which made it impossible to
monkeypatch. Hack around this in the source itself by importing the
query function where it's used instead. Not especially happy with this
solution, I suspect we ought to rework the presto/athena integration
code somehow to make testing easier.