Make measure summary more meaningful
Instead of giving a median of the previous version's counts at the present
time (which is quite biased), calculate the previous version's counts at the
same time as the present one. This helps account for skew as people first
start to update to a previous version.

Also add standard deviation to the API endpoint and web-based summary (to give
at least a superficial idea of whether the current/previous results are
comparable)