Remove newrelic logging of presto query
Turns out that there is a 256 character limit for the size of the field,
which makes the logging not very useful. We could try breaking out
more individual parameters out into their own parameters, but I don't
think it's worth the effort -- my hope was to be able to cut&paste
the field directly from new relic.