Set a stronger limit on the age of the build id in measure etl (partially fixes #230)
It used to be at the absolute minimum of the data range that we're
getting minus the data expiry interval, now it's the latest data
point we've received minus the data expiry interval (which is quite
old).

Testing redash, this reduces the size of the data returned by about 20%,
I'm hoping this enough to fix the issues we're seeing, if not I will adjust
further.