Remove test-ci image from docker-compose configuration
The test-ci image makes it impossible to bring up the docker-compose
environment locally, since it conflicts with the "web" image. Instead,
let's just run the tests inside the web image itself. All of this
is only relevant inside the circle environment.

(note that this change removes the code to call codecov, but that
was not active anyway)